Updated the way bpm is calculated. Currently just takes 8 previous samples, calculates the BPM once, then breaks out of program

+++ button-prototype/button-history.h
@@ -3,8 +3,8 @@
 #include "button.h"
 
 void recordButtonPress(enum buttons button);
-int calculateBPM();
-int getBPM();
+long long calculateBPM();
+long long getBPM();
 void clearHistory();
 void initButtonHistory();
